SongEmotionModel/cloudfunction/requirements.txt
Pillow==7.1.2
torch==1.5.1
torchvision==0.6.1
matplotlib==3.1.1
appdirs==1.4.4
audioread==2.1.9
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
decorator==4.4.2
idna==2.10
joblib==0.17.0
librosa==0.8.0
llvmlite==0.35.0
numba==0.48.0
numpy==1.19.4
packaging==20.7
pooch==1.3.0
pycparser==2.20
pyparsing==2.4.7
requests==2.25.0
resampy==0.2.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
urllib3==1.26.2